HUUUUGE merge from bay - 08.02.16

+++ code/modules/clothing/spacesuits/rig/modules/ninja.dm
@@ -95,7 +95,7 @@
 	playsound(T, "sparks", 50, 1)
 	anim(T,M,'icons/mob/mob.dmi',,"phaseout",,M.dir)
 
-/obj/item/rig_module/teleporter/engage(atom/target)
+/obj/item/rig_module/teleporter/engage(var/atom/target, var/notify_ai)
 
 	if(!..()) return 0
 
@@ -174,6 +174,10 @@
 
 	interface_name = "dead man's switch"
 	interface_desc = "An integrated self-destruct module. When the wearer dies, so does the surrounding area. Do not press this button."
+	var/list/explosion_values = list(1,2,4,5)
+
+/obj/item/rig_module/self_destruct/small
+	explosion_values = list(0,0,3,4)
 
 /obj/item/rig_module/self_destruct/activate()
 	return
@@ -189,17 +193,12 @@
 
 	//OH SHIT.
 	if(holder.wearer.stat == 2)
-		engage()
+		engage(1)
 
-/obj/item/rig_module/self_destruct/engage()
-	explosion(get_turf(src), 1, 2, 4, 5)
-	if(holder && holder.wearer)
-		holder.wearer.drop_from_inventory(src)
-		qdel(holder)
-	qdel(src)
-
-/obj/item/rig_module/self_destruct/small/engage()
-	explosion(get_turf(src), 0, 0, 3, 4)
+/obj/item/rig_module/self_destruct/engage(var/skip_check)
+	if(!skip_check && usr && alert(usr, "Are you sure you want to push that button?", "Self-destruct", "No", "Yes") == "No")
+		return
+	explosion(get_turf(src), explosion_values[1], explosion_values[2], explosion_values[3], explosion_values[4])
 	if(holder && holder.wearer)
 		holder.wearer.drop_from_inventory(src)
 		qdel(holder)
